[2nd week] 3단계 요구사항1 검증 완료
