+++ src/test/java/com/djawnstj/common/User.java
@@ -1,4 +1,4 @@
-package com.djawnstj.mvcframework.bean;
+package com.djawnstj.common;
 
 public class User {
     private final String id;
