+++ src/test/java/com/djawnstj/mvcframework/bean/ComponentScannerTest.java
@@ -1,6 +1,8 @@
 package com.djawnstj.mvcframework.bean;
 
 import com.djawnstj.mvcframework.annotation.Component;
+import com.djawnstj.common.UserRepository;
+import com.djawnstj.common.UserService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -13,11 +15,14 @@ public class ComponentScannerTest {
     @Test
     @DisplayName("특정 애너테이션이 붙은 클래스만 가져오기")
     void getClassInfo() {
-        final String basePackage = "com.djawnstj.mvcframework";
-
+        // Given
+        final String basePackage = "com.djawnstj.common";
         ComponentScanner componentScanner = new ComponentScanner();
+
+        // When
         Set<Class<?>> scanList = componentScanner.scan(basePackage, Component.class);
 
+        // Then
         assertThat(scanList).contains(UserService.class);
         assertThat(scanList).contains(UserRepository.class);
     }
